+++ Libplanet/Blocks/Block.cs
@@ -182,8 +182,8 @@ namespace Libplanet.Blocks
         /// <summary>
         /// Validates the integrity of the <see cref="Block{T}"/>.
         /// <para>It throws an <see cref="InvalidBlockException"/> or
-        /// an <see cref="InvalidTxUpdatedAddressesException"/>
-        /// if there is any integrity error.</para>
+        /// an <see cref="InvalidTxException"/> if there is any
+        /// integrity error.</para>
         /// <para>Otherwise it returns an <see cref="IAccountStateDelta"/>
         /// which represents the final states and maintains the changes
         /// from the states of the previous <see cref="Block{T}"/>.</para>
@@ -212,6 +212,13 @@ namespace Libplanet.Blocks
         /// <exception cref="InvalidBlockNonceException">Thrown when
         /// the <see cref="Nonce"/> does not satisfy its
         /// <see cref="Difficulty"/> level.</exception>
+        /// <exception cref="InvalidTxSignatureException">Thrown when its
+        /// <see cref="Transaction{T}.Signature"/> is invalid or not signed by
+        /// the account who corresponds to its
+        /// <see cref="Transaction{T}.PublicKey"/>.</exception>
+        /// <exception cref="InvalidTxPublicKeyException">Thrown when its
+        /// <see cref="Transaction{T}.Signer"/> is not derived from its
+        /// <see cref="Transaction{T}.PublicKey"/>.</exception>
         /// <exception cref="InvalidTxUpdatedAddressesException">Thrown when
         /// any <see cref="IAction"/> of <see cref="Transactions"/> tries
         /// to update the states of <see cref="Address"/>es not included
@@ -343,6 +350,11 @@ namespace Libplanet.Blocks
                     $"hash ({Hash}) with the nonce ({Nonce}) does not " +
                     $"satisfy its difficulty level {Difficulty}."
                 );
+            }
+
+            foreach (Transaction<T> tx in Transactions)
+            {
+                tx.Validate();
             }
         }